workaround for duplicate title
